fix(plugin-form-builder): run the booting submission suite in the integration lane
The last group in `prepare-submission.test.ts` boots core three times through
`createTestNextly`, but the file is named for the unit lane. The two lanes are
told apart by filename alone: the unit config excludes
`src/**/*.integration.test.ts` and the integration config claims exactly that
suffix. So three instance boots ran under a 10s budget sized for jsdom
component tests, in parallel with the rest of the monorepo, and timed out.

Those three cases move to `prepare-submission.integration.test.ts`, where the
budget is 30s and `fileParallelism` is false. The pure cases stay. No assertion
changed; the imports each file no longer needs are dropped.
